pyproject.toml
[project]
name = "amr-data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.9.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "streamlit>=1.55.0",
]

[dependency-groups]
dev = [
    "pgcli>=4.4.0",
]
